refactor(search): add useSearchState
